requirements.txt
aiogram==3.17.0
python-dotenv==1.0.1

